# docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.AttachmentResponse:
    description: File attachment metadata for boards and projects Contains information
      about uploaded files including S3 URL, size, and content type
    properties:
      contentType:
        example: application/pdf
        type: string
      fileName:
        example: document.pdf
        type: string
      fileSize:
        example: 1024000
        type: integer
      fileUrl:
        example: https://s3.amazonaws.com/bucket/file.pdf
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      uploadedAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      uploadedBy:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
...
  project-board-api_internal_dto.CreateProjectRequest:
    description: Request body for creating a new project with optional start and due
      dates startDate and dueDate are optional, but startDate must be before or equal
      to dueDate if both are provided attachmentIds is an optional array of attachment
      IDs to link to the project
    properties:
      attachmentIds:
        example:
        - f47ac10b-58cc-4372-a567-0e02b2c3d479
        items:
          type: string
        type: array
      description:
        example: Project for launching new product features in Q1 2024
        maxLength: 500
        type: string
      dueDate:
        example: "2024-03-31T23:59:59Z"
        type: string
      name:
        example: Q1 2024 Product Launch
        maxLength: 100
        minLength: 2
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      workspaceId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
    required:
    - name
    - workspaceId
    type: object
...
  project-board-api_internal_dto.FieldTypeInfo:
    properties:
      description:
        type: string
      typeId:
        type: string
      typeName:
        type: string
    type: object
...
  project-board-api_internal_dto.PaginatedProjectsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
        type: array
      total:
        type: integer
    type: object
...
  project-board-api_internal_dto.ProjectInitSettingsResponse:
    properties:
      defaultViewId:
        type: string
      fieldTypes:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.FieldTypeInfo'
        type: array
      fields:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.FieldWithOptionsResponse'
        type: array
      project:
        $ref: '#/definitions/project-board-api_internal_dto.ProjectBasicInfo'
    type: object
...
  project-board-api_internal_dto.ProjectResponse:
    description: Project response with optional start/due dates and attachments startDate
      and dueDate are included only if they were set attachments is an array of file
      metadata (empty array if no attachments)
    properties:
      attachments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.AttachmentResponse'
        type: array
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        example: Project for launching new product features in Q1 2024
        type: string
      dueDate:
        example: "2024-03-31T23:59:59Z"
        type: string
      isPublic:
        example: true
        type: boolean
      name:
        example: Q1 2024 Product Launch
        type: string
      ownerEmail:
        example: owner@example.com
        type: string
      ownerId:
        example: b2c3d4e5-f6a7-8901-bcde-f12345678901
        type: string
      ownerName:
        example: John Doe
        type: string
      projectId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      updatedAt:
        example: "2024-01-15T14:20:00Z"
        type: string
      workspaceId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
...
  project-board-api_internal_dto.UpdateProjectRequest:
    description: Request body for updating a project. All fields are optional. startDate
      must be before or equal to dueDate if both are provided attachmentIds is an
      optional array of attachment IDs to add to the project
    properties:
      attachmentIds:
        example:
        - f47ac10b-58cc-4372-a567-0e02b2c3d479
        items:
          type: string
        type: array
      description:
        example: Updated project description
        maxLength: 500
        type: string
      dueDate:
        example: "2024-04-15T23:59:59Z"
        type: string
      name:
        example: Q1 2024 Product Launch - Updated
        maxLength: 100
        minLength: 2
        type: string
      startDate:
        example: "2024-01-15T00:00:00Z"
        type: string
    type: object
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
    ALB path-based routing with /api/boards prefix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: |-
        특정 Workspace에 속한 모든 Project를 조회합니다. 프론트엔드 호환용 엔드포인트
        각 프로젝트는 startDate, dueDate (설정된 경우), attachments (첨부파일 메타데이터 배열)를 포함합니다
      parameters:
      - description: Workspace ID (UUID)
        in: query
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.PaginatedProjectsResponse'
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 Project 목록 조회 (쿼리 파라미터 방식)
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: |-
        새로운 Project를 생성합니다
        startDate와 dueDate는 선택 사항이며, startDate는 dueDate보다 이전이어야 합니다
      parameters:
      - description: Project 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 생성
      tags:
      - projects
  /projects/{projectId}:
    delete:
      description: Project를 삭제합니다 (OWNER만 가능)
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 삭제 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 삭제
      tags:
      - projects
    get:
      description: |-
        특정 Project의 상세 정보를 조회합니다
        프로젝트는 startDate, dueDate (설정된 경우), attachments (첨부파일 메타데이터 배열)를 포함합니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 상세 조회
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: |-
        Project의 이름, 설명, 날짜를 수정합니다 (OWNER만 가능)
        startDate와 dueDate를 수정할 수 있으며, startDate는 dueDate보다 이전이어야 합니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      - description: Project 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 수정
      tags:
      - projects
...
  /projects/{projectId}/init-settings:
    get:
      description: Project 진입 시 필요한 초기 설정 데이터를 한 번에 조회합니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 초기 설정 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectInitSettingsResponse'
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 초기 설정 조회
      tags:
      - projects
...
  /projects/search:
    get:
      description: Workspace 내에서 Project를 이름이나 설명으로 검색합니다
      parameters:
      - description: Workspace ID (UUID)
        in: query
        name: workspaceId
        required: true
        type: string
      - description: 검색어
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 10
        description: 페이지 크기
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 검색 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.PaginatedProjectsResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 검색
      tags:
      - projects
  /projects/workspace/{workspaceId}:
    get:
      description: |-
        특정 Workspace에 속한 모든 Project를 조회합니다
        각 프로젝트는 startDate, dueDate (설정된 경우), attachments (첨부파일 메타데이터 배열)를 포함합니다
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 Project 목록 조회
      tags:
      - projects
  /projects/workspace/{workspaceId}/default:
    get:
      description: |-
        특정 Workspace의 기본(default) Project를 조회합니다
        프로젝트는 startDate, dueDate (설정된 경우), attachments (첨부파일 메타데이터 배열)를 포함합니다
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 기본 Project 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: 기본 Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 기본 Project 조회
      tags:
      - projects
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"